pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-tooling-index"
version = "1.0.0"
description = "Catalog and analyze Claude Code tooling with TUI dashboard"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Wolfgang Schoenberger"}
]
keywords = ["claude", "claude-code", "tooling", "analytics", "dashboard", "tui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "textual>=0.47.0",
    "rich>=13.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "coverage[toml]>=7.4.0",
    "interrogate>=1.7.0",
    "ruff>=0.6.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
tooling-index = "claude_tooling_index.cli:cli"

[project.urls]
Homepage = "https://github.com/wolfiesch/claude-tooling-index"
Repository = "https://github.com/wolfiesch/claude-tooling-index"
Issues = "https://github.com/wolfiesch/claude-tooling-index/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["claude_tooling_index*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
addopts = [
    "--strict-markers",
    "--strict-config",
]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["claude_tooling_index"]

[tool.coverage.report]
fail_under = 20
show_missing = true
skip_empty = true

[tool.interrogate]
fail-under = 85
verbose = 1
exclude = ["tests", "third_party"]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-module = false
ignore-nested-functions = true
ignore-private = true
omit-covered-files = false
quiet = false

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = [
    "E501",  # Line too long (relaxed for readability)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
